Update base folder definition and refactor file imports
Changed the definition of base folder in 'Package.js' and 'WpPackage.js'. File 'wpConfig.js' was renamed to 'wordpress.js', and all references to this file were updated. Removed redundant import statements and refactored the way functions are imported from utility files. This is done to simplify the code and improve its maintainability. Removed the function 'geVarFromPHPFile' from 'data.js' and moved it to 'wordpress.js'.
