+++ lib/Dump.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const { getCurrentWpInfo, getUserLocale, getLastWp} = require("./utils/wpConfig");
+const { getCurrentWpInfo, getUserLocale, getLastWp } = require("./utils/wordpress.js");
 
 /**
  * Represents a Dump class for WordPress configuration.
@@ -14,7 +14,7 @@ class Dump {
    * Initializes the class with the necessary folders for WordPress.
    */
   constructor (paths) {
-    this.wpFolder = paths.rootFolder;
+    this.wpFolder = paths.baseFolder;
     this.themeFolder = paths.themeFolder;
     this.pluginsFolder = paths.pluginsFolder;
   }
