+++ lib/Initialize.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const {defaultWpConfig, defaultWpInstallFolder, pkgFileName, defaultWpDatabaseType} = require("./constants.js");
-const {getLastWp, getUserLocale} = require("./utils/wpConfig.js");
+const {getLastWp, getUserLocale} = require("./utils/wordpress.js");
 
 /**
  * Initialize class for generating WordPress configuration file.
