+++ lib/Package.js
@@ -30,7 +30,6 @@ class Package {
   constructor (pkgConfig, packageType, paths) {
     this.pkgInfo = pkgConfig;
     this.packageType = packageType;
-    this.rootFolder = paths.rootFolder;
     this.tempDir = paths.tempDir;
     this.destFolder = this.getDestFolder(paths, packageType);
   }
