+++ lib/WpPackage.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const { renameFolder } = require('./utils/fs.js');
 
 const Package = require('./package.js');
-const { replaceDbConstant, replaceDbConstantBool, replaceEmptySalts } = require('./utils/wpConfig.js');
+const { replaceDbConstant, replaceDbConstantBool, replaceEmptySalts } = require('./utils/wordpress.js');
 const { getWordPressDownloadUrl } = require('./utils/data.js');
 
 /**
@@ -23,7 +23,7 @@ class WpPackage extends Package {
     const downloadUrl = getWordPressDownloadUrl(version, language);
 
     try {
-      const destinationPath = path.join(this.rootFolder, this.pkgInfo.name);
+      const destinationPath = path.join(this.baseFolder, this.pkgInfo.name);
 
       if (fs.existsSync(destinationPath)) {
         console.log('🔄️ WordPress folder already exists. Skipping download.');
