+++ lib/index.js
@@ -7,8 +7,8 @@ const Initialize = require('./initialize.js');
 
 const yargs = require('yargs');
 const { hideBin } = require('yargs/helpers');
-const { getWordPressPaths } = require("./utils/wpConfig");
-const Updater = require("./Updater");
+const { getWordPressPaths } = require("./utils/wordpress.js");
+const Updater = require("./Updater.js");
 
 /**
  * @typedef WPMMconfigWP - The configuration object for WordPress
