+++ lib/utils/data.js
@@ -1,26 +1,7 @@
-
 const fs = require('fs');
 const {get} = require("axios");
-const {pkgFileName} = require("../constants");
-const Initialize = require("../Initialize");
-
-/**
- * Reads a PHP file, extracts, and returns the WordPress version number.
- *
- * @param {string} fileContent - Path to the PHP file to read
- * @param variableName - The name of the variable to search adn replace
- * @return {string|null} WordPress version number or null if not found or in case of an error
- */
-function geVarFromPHPFile (fileContent, variableName = '$wp_version') {
-  // Define a regular expression to match the variableName and its value with both the single and double quotes
-  const versionRegex = new RegExp(`${variableName}\\s*=\\s*['"]([^'"]+)['"]`, 'g');
-
-  // Use the regular expression to extract the version number
-  const match = fileContent.match(versionRegex);
-
-  // Return the version number or null if not found
-  return match ? match[1] : null;
-}
+const {pkgFileName} = require("../constants.js");
+const Initialize = require("../Initialize.js");
 
 /**
  * Reads wp-package.json from the root folder and extracts the value of the --template option or uses the default.
@@ -161,7 +142,6 @@ function printTimePassed (startTime) {
 }
 
 module.exports = {
-  geVarFromPHPFile,
   getConfig,
   getInfo,
   getConnectionSettings,
