+++ lib/utils/wordpress.js
@@ -1,9 +1,8 @@
 const path = require("path");
 const fs = require("fs");
-const {geVarFromPHPFile} = require("./data.js");
 const axios = require("axios");
-const {defaultWpInstallFolder} = require("../constants.js");
-const {defaultWpInstallLanguage} = require("../constants");
+const {defaultWpInstallFolder, defaultWpInstallLanguage, pkgFileName} = require("../constants.js");
+
 
 /**
  * Gets the default paths for the WordPress installation.
@@ -12,10 +11,20 @@ const {defaultWpInstallLanguage} = require("../constants");
  * @return {WPMMpaths} - An object containing the default paths for the WordPress installation.
  */
 function getWordPressPaths (config, rootFolder = process.cwd()) {
-  const baseFolder = path.join(rootFolder, config.name ?? defaultWpInstallFolder);
+  /**
+   * the WordPress installation folder
+   *
+   * @type {string}
+   */
+  let baseFolder = rootFolder;
+
+  // if wp-config.php exists in the root folder or the wp-package.json file exists in the root folder
+  if (! fs.existsSync(path.join(rootFolder, 'wp-config.php')) && ! fs.existsSync(path.join(rootFolder, pkgFileName))) {
+    baseFolder = path.join(rootFolder, config.name ?? defaultWpInstallFolder);
+  }
+
   return {
     tempDir: path.join(rootFolder, 'temp'),
-    rootFolder: rootFolder,
     baseFolder: baseFolder,
     pluginsFolder: path.join(baseFolder, 'wp-content', 'plugins'),
     themeFolder: path.join(baseFolder, 'wp-content', 'themes')
@@ -48,6 +57,25 @@ function getUserLocale() {
 }
 
 /**
+ * Reads a PHP file, extracts, and returns the WordPress version number.
+ *
+ * @param {string} fileContent - Path to the PHP file to read
+ * @param variableName - The name of the variable to search adn replace
+ *
+ * @return {string|null} WordPress version number or null if not found or in case of an error
+ */
+function getVarFromPHPFile (fileContent, variableName = '$wp_version') {
+  // Define a regular expression to match the variableName and its value with both the single and double quotes
+  const versionRegex = new RegExp(`${variableName}\\s*=\\s*['"]([^'"]+)['"]`, 'g');
+
+  // Use the regular expression to extract the version number
+  const match = fileContent.match(versionRegex);
+
+  // Return the version number or null if not found
+  return match ? match[1] : null;
+}
+
+/**
  * Retrieves the WordPress version and locale information from a given WordPress folder.
  *
  * @param {string} wpFolder - The path to the WordPress folder.
@@ -58,8 +86,8 @@ function getCurrentWpInfo(wpFolder) {
   // get the WordPress version and the locale from wp-includes/version.php
   const versionFile = path.join(wpFolder, 'wp-includes', 'version.php');
   const versionFileContent = fs.readFileSync(versionFile, 'utf8');
-  const version = geVarFromPHPFile(versionFileContent, 'wp_version');
-  const locale = geVarFromPHPFile(versionFileContent, 'wp_local_package');
+  const version = getVarFromPHPFile(versionFileContent, 'wp_version');
+  const locale = getVarFromPHPFile(versionFileContent, 'wp_local_package');
   return {
     version,
     locale
